Server: Introduce proxy state machine (#9856)
ref pingcap/tiflash#9855, close pingcap/tiflash#9857
