+++ dbms/src/Storages/KVStore/FFI/ProxyFFI.cpp
@@ -94,6 +94,7 @@ EngineStoreApplyRes HandleWriteRaftCmd(const EngineStoreServerWrap * server, Wri
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         return server->tmt->getKVStore()
             ->handleWriteRaftCmd(cmds, header.region_id, header.index, header.term, *server->tmt);
     }
@@ -112,6 +113,7 @@ EngineStoreApplyRes HandleAdminRaftCmd(
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         raft_cmdpb::AdminRequest request;
         raft_cmdpb::AdminResponse response;
         CHECK_PARSE_PB_BUFF(request, req_buff.data, req_buff.len);
@@ -136,6 +138,7 @@ uint8_t NeedFlushData(EngineStoreServerWrap * server, uint64_t region_id)
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto & kvstore = server->tmt->getKVStore();
         return kvstore->needFlushRegionData(region_id, *server->tmt);
     }
@@ -157,6 +160,7 @@ uint8_t TryFlushData(
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto & kvstore = server->tmt->getKVStore();
         return kvstore->tryFlushRegionData(
             region_id,
@@ -286,6 +290,7 @@ void HandleConsumeWriteBatch(const EngineStoreServerWrap * server, RawVoidPtr pt
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto uni_ps = server->tmt->getContext().getWriteNodePageStorage();
         auto * wb = reinterpret_cast<UniversalWriteBatch *>(ptr);
         LOG_TRACE(&Poco::Logger::get("ProxyFFI"), "FFI consume write batch {}", wb->toString());
@@ -303,6 +308,7 @@ CppStrWithView HandleReadPage(const EngineStoreServerWrap * server, BaseBuffView
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto uni_ps = server->tmt->getContext().getWriteNodePageStorage();
         RaftDataReader reader(*uni_ps);
         auto * page = new Page(reader.read(UniversalPageId(page_id.data, page_id.len)));
@@ -337,6 +343,7 @@ RawCppPtrCarr HandleScanPage(const EngineStoreServerWrap * server, BaseBuffView 
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         LOG_TRACE(
             &Poco::Logger::get("ProxyFFI"),
             "FFI scan page from {} to {}",
@@ -383,6 +390,7 @@ CppStrWithView HandleGetLowerBound(const EngineStoreServerWrap * server, BaseBuf
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto uni_ps = server->tmt->getContext().getWriteNodePageStorage();
         RaftDataReader reader(*uni_ps);
         auto page_id_opt = reader.getLowerBound(UniversalPageId(raw_page_id.data, raw_page_id.len));
@@ -418,6 +426,7 @@ uint8_t IsPSEmpty(const EngineStoreServerWrap * server)
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto uni_ps = server->tmt->getContext().getWriteNodePageStorage();
         return uni_ps->isEmpty();
     }
@@ -432,6 +441,7 @@ void HandlePurgePageStorage(const EngineStoreServerWrap * server)
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto uni_ps = server->tmt->getContext().getWriteNodePageStorage();
         uni_ps->gc({});
     }
@@ -460,6 +470,7 @@ void HandleDestroy(EngineStoreServerWrap * server, uint64_t region_id)
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto & kvstore = server->tmt->getKVStore();
         Stopwatch watch;
         kvstore->handleDestroy(region_id, *server->tmt);
@@ -476,6 +487,7 @@ EngineStoreApplyRes HandleIngestSST(EngineStoreServerWrap * server, SSTViewVec s
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto & kvstore = server->tmt->getKVStore();
         return kvstore->handleIngestSST(header.region_id, snaps, header.index, header.term, *server->tmt);
     }
@@ -492,6 +504,7 @@ StoreStats HandleComputeStoreStats(EngineStoreServerWrap * server)
     StoreStats res{}; // res.fs_stats.ok = false by default
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto global_capacity = server->tmt->getContext().getPathCapacity();
         res.fs_stats = global_capacity->getFsStats();
         // TODO: set engine read/write stats
@@ -661,6 +674,7 @@ RawCppPtr PreHandleSnapshot(
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         metapb::Region region;
         CHECK_PARSE_PB_BUFF(region, region_buff.data, region_buff.len);
         auto & tmt = *server->tmt;
@@ -697,6 +711,7 @@ void ApplyPreHandledSnapshot(EngineStoreServerWrap * server, RawVoidPtr res, Raw
         auto * snap = reinterpret_cast<PreHandledSnapshotWithFiles *>(res);
         try
         {
+            RUNTIME_CHECK(server->tmt != nullptr);
             auto & kvstore = server->tmt->getKVStore();
             kvstore->applyPreHandledSnapshot(
                 RegionPtrWithSnapshotFiles{snap->region, std::move(snap->prehandle_result.ingest_ids)},
@@ -724,6 +739,7 @@ void AbortPreHandledSnapshot(EngineStoreServerWrap * server, uint64_t region_id,
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         UNUSED(peer_id);
         auto & kvstore = server->tmt->getKVStore();
         kvstore->abortPreHandleSnapshot(region_id, *server->tmt);
@@ -746,6 +762,7 @@ void ReleasePreHandledSnapshot(EngineStoreServerWrap * server, RawVoidPtr res, R
     auto * snap = reinterpret_cast<PreHandledSnapshotWithFiles *>(res);
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto s = RegionPtrWithSnapshotFiles{snap->region, std::move(snap->prehandle_result.ingest_ids)};
         auto & kvstore = server->tmt->getKVStore();
         kvstore->releasePreHandledSnapshot(s, *server->tmt);
@@ -761,6 +778,7 @@ bool KvstoreRegionExists(EngineStoreServerWrap * server, uint64_t region_id)
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto & kvstore = server->tmt->getKVStore();
         return kvstore->getRegion(region_id) != nullptr;
     }
@@ -880,6 +898,7 @@ CppStrWithView GetConfig(EngineStoreServerWrap * server, [[maybe_unused]] uint8_
     std::string config_file_path;
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         config_file_path = server->tmt->getContext().getConfigRef().getString("config-file");
         std::ifstream stream(config_file_path);
         if (!stream)
@@ -907,7 +926,7 @@ void SetStore(EngineStoreServerWrap * server, BaseBuffView buff)
     metapb::Store store{};
     CHECK_PARSE_PB_BUFF(store, buff.data, buff.len);
     assert(server);
-    assert(server->tmt);
+    RUNTIME_CHECK(server->tmt != nullptr);
     assert(store.id() != 0);
     server->tmt->getKVStore()->setStore(std::move(store));
 }
@@ -977,6 +996,7 @@ void HandleSafeTSUpdate(
 {
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         RegionTable & region_table = server->tmt->getRegionTable();
         region_table.updateSafeTS(region_id, leader_safe_ts, self_safe_ts);
     }
@@ -1003,6 +1023,7 @@ BaseBuffView GetLockByKey(const EngineStoreServerWrap * server, uint64_t region_
     auto tikv_key = TiKVKey(key.data, key.len);
     try
     {
+        RUNTIME_CHECK(server->tmt != nullptr);
         auto & kvstore = server->tmt->getKVStore();
         auto region = kvstore->getRegion(region_id);
         auto value = region->getLockByKey(tikv_key);
